+++ src/lib/evolution.ts
@@ -75,7 +75,7 @@ function add_history(evo_deleted) {
     const parent_item = get_item(
       evo.parent_entity,
       evo.parent_entity_id,
-      evo_deleted
+      evo_deleted,
     )
     evo.parent_name = parent_item?.name
     evo.parent_deleted = parent_item?._deleted
@@ -109,17 +109,17 @@ function add_validity(validities, type, entity, entity_data, evo_deleted) {
   const parent_item = get_item(
     parent_entity,
     entity_data[`${parent_entities[entity]}_id`],
-    evo_deleted
+    evo_deleted,
   )
 
   const timestamp = date_to_timestamp(
     entity_data[type],
-    type === 'start_date' ? 'start' : 'end'
+    type === 'start_date' ? 'start' : 'end',
   )
 
   if (!timestamp) {
     console.error(
-      `Invalid date format for ${type} in ${entity} with id ${entity_data.id}, value = ${entity_data[type]}`
+      `Invalid date format for ${type} in ${entity} with id ${entity_data.id}, value = ${entity_data[type]}`,
     )
     return
   }
@@ -173,7 +173,7 @@ function add_validities(evo_deleted) {
             'start_date',
             entity,
             entity_data,
-            evo_deleted
+            evo_deleted,
           )
         }
         if ('end_date' in entity_data && entity_data.end_date) {
@@ -185,7 +185,7 @@ function add_validities(evo_deleted) {
             'last_update_date',
             entity,
             entity_data,
-            evo_deleted
+            evo_deleted,
           )
         }
         if ('next_update_date' in entity_data && entity_data.next_update_date) {
@@ -194,7 +194,7 @@ function add_validities(evo_deleted) {
             'next_update_date',
             entity,
             entity_data,
-            evo_deleted
+            evo_deleted,
           )
         }
       })
@@ -235,8 +235,8 @@ function output_diff_number(oldVal, newVal) {
 
   return `${oldVal.toLocaleString()} ${arrow_right} ${newVal.toLocaleString()} 
       <br><span class="${diffClass}">${
-    diff > 0 ? '+' : ''
-  }${diff.toLocaleString()} | 
+        diff > 0 ? '+' : ''
+      }${diff.toLocaleString()} | 
       ${diff > 0 ? '+' : ''}${percentageChange}%</span>
   `
 }