+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - add: community standards docs (code of conduct, security, issue templates, pr templates)
 - add: code quality check on ci
+- add: prettier check and format all code with it
 - change: rename all ts files, image and folder to be in kebab-case
 - change: json config files use camelCase instead of snake_case
 - fix: fix eslint path, ci ignore path and contributing.md